+++ YubiKit/YubiKit/YubiKit.h
@@ -67,6 +67,7 @@
 #import "YKFU2FRegisterResponse.h"
 
 #import "YKFPCSC.h"
+#import "YKFPCSCLayer.h"
 
 #import "YKFNSDataAdditions.h"
 #import "YKFWebAuthnClientData.h"